Adjusting the static utility classes.
Renaming static utility classes api, helpers, and prolog to conform to Java conventions.
Giving each of those classes a private constructor.
Providing the Api class with a static instance of UnknownCyberFileProvider to avoid requiring it in every wrapped api call.
Include in the Api class a getter and setter method for the fileProvider static instance.